Completed merge of MeasurePointTool.
